Projeto I/python_etl/requirements.txt
pandas==1.3.3
psycopg2-binary==2.9.1
pymongo==3.12.0

